+++ multi-step-form/app/_components/StepContent.tsx
@@ -20,8 +20,6 @@ function StepContent({stepNumber, setStepNumber, formData }: StepContentProps): 
 
     const [values, setValues] = useState(initialValues)
 
-    console.log(values)
-
     if(stepNumber === formData.steps.length+2) return <Confirmation/>
     if(stepNumber === formData.steps.length+1) return <Summary stepNumber={stepNumber} formData={formData} setStepNumber={setStepNumber} values={values}/>
