+++ multi-step-form/app/_components/steps/MultipleSelect.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import Button from "@/app/_components/Button";
-import AddOnSelector from "@/app/_components/AddOnSelector";
+import MultipleSelectOption from "@/app/_components/MultipleSelectOption";
 import {InitialValuesInterface, multipleSelectOption} from "@/app/interfaces"
 import NavigationButtons from "@/app/_components/NavigationButtons";
 
@@ -18,8 +17,8 @@ function MultipleSelect({setValues, setStepNumber, values, fields, stepNumber} :
         <div className="flex flex-col my-4 h-full justify-between">
             <div className="flex flex-col gap-8">
                 <div className="flex flex-col gap-4 justify-between">
-                    {fields.map((option: multipleSelectOption, i: number) => (<AddOnSelector index={i} key={option.optionTitle} title={option.optionTitle}
-                                                                  primary={option.primary} description={option.optionDescription} setValues={setValues} values={values} stepNumber={stepNumber}/>))}
+                    {fields.map((option: multipleSelectOption, i: number) => (<MultipleSelectOption index={i} key={option.optionTitle} title={option.optionTitle}
+                                                                                                    primary={option.primary} description={option.optionDescription} setValues={setValues} values={values} stepNumber={stepNumber}/>))}
                 </div>
             </div>
             <NavigationButtons step={stepNumber} setStepNumber={setStepNumber}/>
